chore: move hono to new integrations folder
