+++ packages/galaxy/Dockerfile
@@ -20,7 +20,7 @@ RUN chown node:node /app
 # Copy root node modules and any utilized packages
 COPY --from=builder /app/node_modules /app/node_modules
 COPY --from=builder /app/packages/mock-server /app/packages/mock-server
-COPY --from=builder /app/packages/hono-api-reference /app/packages/hono-api-reference
+COPY --from=builder /app/integrations/hono-api-reference /app/integrations/hono-api-reference
 COPY --from=builder /app/packages/oas-utils /app/packages/oas-utils
 COPY --from=builder /app/packages/openapi-parser /app/packages/openapi-parser
 COPY --from=builder /app/packages/galaxy /app/packages/galaxy
