+++ packages/galaxy/playground/index.ts
@@ -1,7 +1,8 @@
 import { serve } from '@hono/node-server'
-import { apiReference } from '@scalar/hono-api-reference'
 import { createMockServer } from '@scalar/mock-server'
 import fs from 'fs/promises'
+
+import { apiReference } from '../../../integrations/hono-api-reference/src'
 
 const specification = await readOpenApiDocumentFromDisk()
