+++ tooling/cloudbuild/cloudbuildEnv.json
@@ -17,7 +17,7 @@
   {
     "name": "hono",
     "cloudBuildName": "_SERVICE_HONO",
-    "dockerPath": "./packages/hono-api-reference/Dockerfile"
+    "dockerPath": "./integrations/hono-api-reference/Dockerfile"
   },
   {
     "name": "nest-js-express",
